+++ include/optic_flow/utilityFunctions.h
@@ -8,46 +8,46 @@
 struct StatData
 {
   uint  num;
-  float mean;
-  float stdDev;
+  double mean;
+  double stdDev;
 
-  float meanX;
-  float stdDevX;
+  double meanX;
+  double stdDevX;
 
-  float meanY;
-  float stdDevY;
+  double meanY;
+  double stdDevY;
 };
 
 struct SpeedBox
 {
   ros::Time   time;
-  cv::Point2f speed;
-  cv::Point2f odometry_speed;
+  cv::Point2d speed;
+  cv::Point2d odometry_speed;
 };
 
 void                     rotate2d(double &x, double &y, double alpha);
-void                     rotate2d(cv::Point2f &pt, double alpha);
-cv::Point2f              pointMean(std::vector<cv::Point2f> pts);
-float                    getDistSq(cv::Point2f p1, cv::Point2f p2);
-float                    getNormSq(cv::Point2f p1);
-cv::Point2f              twoPointMean(cv::Point2f p1, cv::Point2f p2);
+void                     rotate2d(cv::Point2d &pt, double alpha);
+cv::Point2d              pointMean(std::vector<cv::Point2d> pts);
+double                    getDistSq(cv::Point2d p1, cv::Point2d p2);
+double                    getNormSq(cv::Point2d p1);
+cv::Point2d              twoPointMean(cv::Point2d p1, cv::Point2d p2);
 double                   calcMean(std::vector<double> pts);
-cv::Point2f              allsacMean(std::vector<cv::Point2f> pts, float thresholdRadius_sq, int *chosen);
-double                   allsacMean(std::vector<double> pts, float thresholdRadius, int *chosen);
-void                     multiplyAllPts(std::vector<cv::Point2f> &v, float mulx, float muly);
-void                     multiplyAllPts(std::vector<double> &v, float mul);
-void                     rotateAllPts(std::vector<cv::Point2f> &v, double alpha);
-void                     addToAll(std::vector<cv::Point2f> &v, float adx, float ady);
-cv::Point2f              ransacMean(std::vector<cv::Point2f> pts, int numOfChosen, float thresholdRadius_sq, int numOfIterations);
-std::vector<cv::Point2f> getOnlyInAbsBound(std::vector<cv::Point2f> v, float up);
+cv::Point2d              allsacMean(std::vector<cv::Point2d> pts, double thresholdRadius_sq, int *chosen);
+double                   allsacMean(std::vector<double> pts, double thresholdRadius, int *chosen);
+void                     multiplyAllPts(std::vector<cv::Point2d> &v, double mulx, double muly);
+void                     multiplyAllPts(std::vector<double> &v, double mul);
+void                     rotateAllPts(std::vector<cv::Point2d> &v, double alpha);
+void                     addToAll(std::vector<cv::Point2d> &v, double adx, double ady);
+cv::Point2d              ransacMean(std::vector<cv::Point2d> pts, int numOfChosen, double thresholdRadius_sq, int numOfIterations);
+std::vector<cv::Point2d> getOnlyInAbsBound(std::vector<cv::Point2d> v, double up);
 std::vector<double>      getOnlyInAbsBound(std::vector<double> v, double up);
-std::vector<cv::Point2f> removeNanPoints(std::vector<cv::Point2f> v);
+std::vector<cv::Point2d> removeNanPoints(std::vector<cv::Point2d> v);
 std::vector<double>      removeNanPoints(std::vector<double> v);
-std::vector<cv::Point2f> getOnlyInRadiusFromTruth(cv::Point2f truth, std::vector<cv::Point2f> v, float rad);
-float                    absf(float x);
+std::vector<cv::Point2d> getOnlyInRadiusFromTruth(cv::Point2d truth, std::vector<cv::Point2d> v, double rad);
+double                    absf(double x);
 double                   absd(double x);
 StatData                 analyzeSpeeds(ros::Time fromTime, std::vector<SpeedBox> speeds);
-std::vector<cv::Point2f> estimateTranRotVvel(std::vector<cv::Point2f> vectors, double a, double fx, double fy, double range, double allsac_radius,
+std::vector<cv::Point2d> estimateTranRotVvel(std::vector<cv::Point2d> vectors, double a, double fx, double fy, double range, double allsac_radius,
                                              double duration, double max_vert_speed, double max_yaw_speed);
 
 #endif  // UTILITYFUNCTIONS_H
